igc: Fix fail to set fragment size via ethtool
BugLink: https://bugs.launchpad.net/bugs/1945461

Fix regression which fail to set fragment size from 256 bytes to 260 bytes
due to data type ranges. Changes the data type range to u32.

(cherry picked from commit 7cc95dd02afe42f37fae2f3332b5e3731e5e0dbc/linux-intel-5.13)

+++ drivers/net/ethernet/intel/igc/igc.h
@@ -187,7 +187,7 @@ struct igc_adapter {
 	ktime_t cycle_time;
 	bool qbv_enable;
 	bool frame_preemption_active;
-	u8 add_frag_size;
+	u32 add_frag_size;
 
 	/* OS defined structs */
 	struct pci_dev *pdev;
